Refactor dependencies into class
While working on the dead process restart feature, I ran into a couple
problems:

* `matcher_registry` imports were winding up everywhere in order to get
  matchers to rerun match processes. This lead to circular imports all
  over as well.
* I was having to pass in storage, workspace, pickelable deps, and matcher
  dependencies to some of the new methods, which was making the method
  signatures huge.

To fix this I refactored dependency handling into a state class that can be
passed around and used to get dependencies as needed. I added the matchers
to that class so that the matcher registry only needs to be imported at the
top level of the app and the matchers passed into the state builder, which
breaks the import loops.