Send process heartbeat immediately
This reduces the period when multiple processes can be started for
short jobs that never send a heartbeat after the creation time is
> 60s in the past.

To reduce further, could add a process started field that the main
process updates when it starts the process and take that into
account in the restart calculation. This change should fix most of
the problem, though, so don't worry about that for now.