Move and rename init_storage script
Will also be used for creating and updating views, so will get quite a
bit more complex and be used more often, so putting in a more obvious
place. Will update documentation once the script is updated. This commit
just moves the script to avoid mixing a file move and substantial file
changes in the same commit.